feat: Add translations for language change dialog
This commit adds new string resources for the "Change language" dialog to provide translations for multiple languages.

The new strings include the dialog title, a message informing the user that the app will restart, and the positive ("Restart") and negative ("Cancel") action buttons.

Translations have been added for the following languages:
- Catalan (ca)
- German (de)
- Greek (el)
- Spanish (es)
- Basque (eu)
- Persian (fa)
- French (fr)
- Hebrew (iw)
- Italian (it)
- Japanese (ja)
- Polish (pl)
- Brazilian Portuguese (pt-rBR)
- Russian (ru)
- Turkish (tr)
- Vietnamese (vi)
- Chinese (zh)
- Traditional Chinese (zh-rTW)